bandwidth: work-around for bug in c3js (tab inactive)
c3js/c3#1097
This is a temporary work-around. It should be removed upon bugfix in
c3js library.
When flow() method is called and tab is inactive in the browser, points
of the past are not popped properly. They accumulate and cause the browser
to freeze completely.
The trick is to defer the call to flow() until the tab is active.
For this we need to keep track of the points in a buffer when tab is
inactive.