Prevent luci apps from being upgraded all the time (#216)
luci-app-overthebox, luci-base and luci-mod-admin-full were upgraded at
each OTB upgrade, even if those packets were already up-to-date. This is
because by default, if PKG_VERSION is not set, luci generates a
PKG_VERSION using the git commit sha1 and the date. This caused the
version number to change all the time, triggering useless upgrades each
time.

https://github.com/ovh/overthebox-feeds/blob/master/luci/luci.mk#L61

Signed-off-by: Martin Wetterwald <martin.wetterwald@corp.ovh.com>