ss-redir-init: Be more tolerant in uci syntax
In the ss-redir init script, we fill the iptables chain "socks_hook"
with one rule per interface belonging to the zone "lan".
Sometimes, the syntax used in the uci configuration is:
    list network 'lan1'
    list network 'lan2'
whereas sometimes, it's:
    option network 'lan1 lan2'

We only supported the former syntax. This patch add support for the
latter syntax to me more robust.