swconfig: Implement smarter timeout when receiving
A timeout can be specified when a command is to be sent.
This timeout will only apply when waiting for the first char of the
response of that command (real response after our own command echo).
This gives time to the Switch CPU to process the user request before it
starts answering.