overthebox: Setup MSS clamping in lan
Some users experience hangs while using the port forwarding from the
public IP to a specific web server in the lan. If the web server
response is to big, the OTB will send a ICMP packet to tell it to
defragment the packet. If the web server does not handle this
fragmentation the end user will never receive the data.
Enabling the MSS clamping on the lan will allow both parties to
negotiate the correct MSS and to communicate without fragmentation.